+++ packages/components/src/Layout/Grid/Grid.tsx
@@ -24,9 +24,21 @@
 
  */
 
-export { Box, BoxProps } from './Box'
-export { Flex, FlexProps } from './Flex'
-export { FlexItem, FlexItemProps } from './FlexItem'
+import styled from 'styled-components'
+import { defaultSpaceSize, SpaceHelperProps } from '../Space'
+import { simpleLayoutCSS } from '../utils/simple'
 
-export * from './Space'
-export * from './Grid'
+export interface GridProps extends SpaceHelperProps {
+  /**
+   * @default 2
+   */
+  columns?: number
+}
+
+export const Grid = styled.div<GridProps>`
+  ${simpleLayoutCSS}
+
+  display: grid;
+  grid-template-columns: repeat(${({ columns }) => columns || 2}, 1fr);
+  grid-gap: ${({ gap, theme }) => theme.space[gap || defaultSpaceSize]};
+`