+++ packages/components/src/Layout/Grid/index.ts
@@ -23,10 +23,4 @@
  SOFTWARE.
 
  */
-
-export { Box, BoxProps } from './Box'
-export { Flex, FlexProps } from './Flex'
-export { FlexItem, FlexItemProps } from './FlexItem'
-
-export * from './Space'
 export * from './Grid'