+++ packages/components/src/Layout/Space/index.ts
@@ -24,9 +24,5 @@
 
  */
 
-export { Box, BoxProps } from './Box'
-export { Flex, FlexProps } from './Flex'
-export { FlexItem, FlexItemProps } from './FlexItem'
-
 export * from './Space'
-export * from './Grid'
+export * from './SpaceVertical'