+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Layout components
+  - `Grid` component
+  - `Space` & `SpaceVertical` components
 - `SelectMulti` component
 - `CodeBlock` component
 - `InputTimeSelect` component